Moved client to own assembly and made it work
