+++ conway.sln
@@ -18,6 +18,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "conwaywebapi", "conwaywebap
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "conwaywebmodel", "conwaywebmodel\conwaywebmodel.csproj", "{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "conwaywebclient", "conwaywebclient\conwaywebclient.csproj", "{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -88,6 +90,18 @@ Global
 		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Release|x64.Build.0 = Release|Any CPU
 		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Release|x86.ActiveCfg = Release|Any CPU
 		{89C41C49-B0B7-4BAE-A7B9-284E416E1CE5}.Release|x86.Build.0 = Release|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Debug|x64.Build.0 = Debug|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Debug|x86.Build.0 = Debug|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Release|x64.ActiveCfg = Release|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Release|x64.Build.0 = Release|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Release|x86.ActiveCfg = Release|Any CPU
+		{9D2A6069-C63E-44E2-86B8-71E11DC9CC77}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
