+++ conwaywebclient/conwaywebclient.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <ProjectReference Include="..\conwaywebmodel\conwaywebmodel.csproj" />
+    <ProjectReference Include="..\conwaylib\conwaylib.csproj" />
+  </ItemGroup>
+  <ItemGroup>   
+    <PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="5.2.9" />
+  </ItemGroup>
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+    <RootNamespace>ConwayWebClient</RootNamespace>
+  </PropertyGroup>
+
+</Project>
